requirements.txt
# nQ-Swap Alpha Bot — Python dependencies
# Install with: pip install -r requirements.txt

# Discord bot framework
discord.py>=2.4.0

# Async HTTP client (replaces blocking requests)
aiohttp>=3.9.0

# Environment variable management
python-dotenv>=1.0.0
